break run_on_slurm into slurm.py and run.py, update README
